Updated project structure and ContactPage
Made modifications that doesn't completely take away from the original
but rather makes some improvements and minor changes.

Removed the nested source code and placed as root directory.

Updated formatting on HTML.
  - Added `main` element to contain all the main content of the page
  - Used indenting to make it easier to read where necessary
  - Wrapped label/input groups into a div for organization and styling

Fixed script tag javascript, which had missing brackets and incorrect
syntax. I believe it was meant to be an IIFE so I fixed it to be that.

Updated styling to include an imported font, "Fredoka" and other
adjustments.

Added `normalize.css` file.