Update Homepage
Minor edits to HTML.

Added the "Fredoka" font.

Modified the header to be uniform across all pages.

Modified the CSS so that the image is an its' own element within the
content and not a background-image for the section element which can be
used in various areas that wouldn't want the image to also be used.

Removed unecessary padding and margins and make the background image
span the entire initial viewport, which is what I believe they wanted to
do.

Centered the text and reduced font-weight.